+++ server/server-agent-puppeteer.js
@@ -24,8 +24,8 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
     function step(op) {
         if (f) throw new TypeError("Generator is already executing.");
         while (_) try {
-            if (f = 1, y && (t = y[op[0] & 2 ? "return" : op[0] ? "throw" : "next"]) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [0, t.value];
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
             switch (op[0]) {
                 case 0: case 1: t = op; break;
                 case 4: _.label++; return { value: op[1], done: false };
@@ -84,7 +84,7 @@ var ServerAgent;
     }(server_agent_js_1.ServerAgent.BaseServer));
     var args = sysProcess.argv.slice();
     args.splice(0, 1);
-    var server = new PuppeteerServerAgent('puppeteer', args, sysOs.platform(), sysProcess.cwd(), sysFs.existsSync, sysFs.unlinkSync, fsix_js_1.fsix.mkdirpSync, fsix_js_1.fsix.readUtf8Sync, sysFs.writeFileSync, sysPath.posix.join);
+    var server = new PuppeteerServerAgent('puppeteer', args, sysOs.platform(), sysProcess.cwd(), sysFs.existsSync, function (fileName) { return sysFs.statSync(fileName).isDirectory(); }, sysFs.unlinkSync, fsix_js_1.fsix.mkdirpSync, fsix_js_1.fsix.readUtf8Sync, sysFs.writeFileSync, sysPath.posix.join);
     server.start();
     // ------------------------------------------------------------------------
     //                               Run Server