+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.12] -
+### Fixed
+- Fixed `abeamer create .` and `abeamer create`.
+- Fixed command line error messages.
+### Added
+- Added extra file and directory exists tests to server-agent.
+- Added more examples to the command line documentation.
+
+
 ## [0.2.11] -
 ### Changed
 - Improved `README.md`.